move NegStatus from neg to annotation
