+++ pkg/annotations/service_test.go
@@ -18,7 +18,6 @@ package annotations
 
 import (
 	"fmt"
-	"k8s.io/ingress-gce/pkg/neg/types"
 	"reflect"
 	"testing"
 
@@ -172,7 +171,7 @@ func TestNEGStatus(t *testing.T) {
 	for _, tc := range []struct {
 		desc            string
 		svc             *v1.Service
-		expectNegStatus *types.NegStatus
+		expectNegStatus *NegStatus
 		expectError     error
 		expectFound     bool
 	}{
@@ -191,7 +190,7 @@ func TestNEGStatus(t *testing.T) {
 					},
 				},
 			},
-			expectNegStatus: &types.NegStatus{NetworkEndpointGroups: types.PortNegMap{"80": "neg-name"}, Zones: []string{"us-central1-a"}},
+			expectNegStatus: &NegStatus{NetworkEndpointGroups: PortNegMap{"80": "neg-name"}, Zones: []string{"us-central1-a"}},
 			expectFound:     true,
 			expectError:     nil,
 		},
@@ -406,5 +405,57 @@ func TestBackendConfigs(t *testing.T) {
 		if !reflect.DeepEqual(configs, tc.expectedConfigs) || tc.expectedErr != err {
 			t.Errorf("%s: for annotations %+v; svc.GetBackendConfigs() = %v, %v; want %v, %v", tc.desc, svc.v, configs, err, tc.expectedConfigs, tc.expectedErr)
 		}
+	}
+}
+
+func TestParseNegStatus(t *testing.T) {
+	for _, tc := range []struct {
+		desc            string
+		status          string
+		expectNegStatus *NegStatus
+		expectError     error
+	}{
+		{
+			desc:            "Test empty string",
+			status:          "",
+			expectNegStatus: &NegStatus{},
+			expectError:     fmt.Errorf("unexpected end of JSON input"),
+		},
+		{
+			desc:            "Test basic status",
+			status:          `{"network_endpoint_groups":{"80":"neg-name"},"zones":["us-central1-a"]}`,
+			expectNegStatus: &NegStatus{NetworkEndpointGroups: PortNegMap{"80": "neg-name"}, Zones: []string{"us-central1-a"}},
+			expectError:     nil,
+		},
+		{
+			desc:            "Test NEG status with 2 ports",
+			status:          `{"network_endpoint_groups":{"80":"neg-name", "443":"another-neg-name"},"zones":["us-central1-a"]}`,
+			expectNegStatus: &NegStatus{NetworkEndpointGroups: PortNegMap{"80": "neg-name", "443": "another-neg-name"}, Zones: []string{"us-central1-a"}},
+			expectError:     nil,
+		},
+		{
+			desc:            "Incorrect fields",
+			status:          `{"network_endpoint_group":{"80":"neg-name"},"zone":["us-central1-a"]}`,
+			expectNegStatus: &NegStatus{},
+			expectError:     nil,
+		},
+		{
+			desc:            "Invalid annotation",
+			status:          `{"network_endpoint_groups":{"80":"neg-name"},"zones":"us-central1-a"]}`,
+			expectNegStatus: &NegStatus{},
+			expectError:     fmt.Errorf("invalid character ']' after object key:value pair"),
+		},
+	} {
+		t.Run(tc.desc, func(t *testing.T) {
+			negStatus, err := ParseNegStatus(tc.status)
+
+			if fmt.Sprintf("%q", err) != fmt.Sprintf("%q", tc.expectError) {
+				t.Errorf("Test case %q: Expect error to be %q, but got: %q", tc.desc, tc.expectError, err)
+			}
+
+			if !reflect.DeepEqual(*tc.expectNegStatus, negStatus) {
+				t.Errorf("Expected NegStatus to be %v, got %v instead", tc.expectNegStatus, negStatus)
+			}
+		})
 	}
 }
